renamed module to encryption-utils
