+++ common/encryption-utils/src/main/java/com/mmadu/encryption/DefaultPropertiesMasterKeyResolver.java
@@ -1,0 +1,30 @@
+package com.mmadu.encryption;
+
+import org.springframework.util.StringUtils;
+
+public class DefaultPropertiesMasterKeyResolver implements MasterKeyResolver {
+    private String defaultMasterKey;
+    private MasterKeyGenerator masterKeyGenerator;
+
+    public void setDefaultMasterKey(String defaultMasterKey) {
+        this.defaultMasterKey = defaultMasterKey;
+    }
+
+    public void setMasterKeyGenerator(MasterKeyGenerator masterKeyGenerator) {
+        this.masterKeyGenerator = masterKeyGenerator;
+    }
+
+    @Override
+    public String getMasterKey() {
+        if (StringUtils.isEmpty(defaultMasterKey)) {
+            defaultMasterKey = generateMasterKey();
+        }
+        return defaultMasterKey;
+    }
+
+    private String generateMasterKey() {
+        String masterKey = masterKeyGenerator.generateMasterKey();
+        System.out.println("Generated master key: " + masterKey);
+        return masterKey;
+    }
+}
