+++ common/encryption-utils/src/main/java/com/mmadu/encryption/KeyCipher.java
@@ -1,0 +1,8 @@
+package com.mmadu.encryption;
+
+public interface KeyCipher {
+
+    String encrypt(String data);
+
+    String decrypt(String encryptedData);
+}
