+++ common/encryption-utils/src/main/java/com/mmadu/encryption/MasterKeyGenerator.java
@@ -1,0 +1,5 @@
+package com.mmadu.encryption;
+
+public interface MasterKeyGenerator {
+    String generateMasterKey();
+}
