+++ common/encryption-utils/src/main/java/com/mmadu/encryption/MasterKeyResolver.java
@@ -1,0 +1,5 @@
+package com.mmadu.encryption;
+
+public interface MasterKeyResolver {
+    String getMasterKey();
+}
