+++ common/pom.xml
@@ -7,9 +7,10 @@
     <groupId>com.mmadu</groupId>
     <artifactId>common</artifactId>
     <version>1.2.1-SNAPSHOT</version>
+    <packaging>pom</packaging>
 
     <modules>
-        <module>security-utils</module>
+        <module>encryption-utils</module>
     </modules>
 
 </project>
