--- common/security-utils/src/main/java/com/mmadu/security/KeyCipher.java
@@ -1,8 +1,0 @@
-package com.mmadu.security;
-
-public interface KeyCipher {
-
-    String encrypt(String data);
-
-    String decrypt(String encryptedData);
-}
