--- common/security-utils/src/main/java/com/mmadu/security/MasterKeyGenerator.java
@@ -1,5 +1,0 @@
-package com.mmadu.security;
-
-public interface MasterKeyGenerator {
-    String generateMasterKey();
-}
