--- common/security-utils/src/main/java/com/mmadu/security/MasterKeyResolver.java
@@ -1,5 +1,0 @@
-package com.mmadu.security;
-
-public interface MasterKeyResolver {
-    String getMasterKey();
-}
