--- common/security-utils/src/main/java/com/mmadu/security/NoOpKeyCipher.java
@@ -1,14 +1,0 @@
-package com.mmadu.security;
-
-public class NoOpKeyCipher implements KeyCipher {
-
-    @Override
-    public String encrypt(String data) {
-        return data;
-    }
-
-    @Override
-    public String decrypt(String encryptedData) {
-        return encryptedData;
-    }
-}
