+++ mmadu-user-service/pom.xml
@@ -99,6 +99,11 @@
 			<artifactId>kweeri-mongo</artifactId>
 			<version>0.0.1</version>
 		</dependency>
+		<dependency>
+			<groupId>com.mmadu</groupId>
+			<artifactId>encryption-utils</artifactId>
+			<version>1.2.1-SNAPSHOT</version>
+		</dependency>
 	</dependencies>
 
 	<build>
