--- mmadu-user-service/src/main/java/com/mmadu/service/providers/KeyCipher.java
@@ -1,8 +1,0 @@
-package com.mmadu.service.providers;
-
-public interface KeyCipher {
-
-    String encrypt(String data);
-
-    String decrypt(String encryptedData);
-}
