--- mmadu-user-service/src/main/java/com/mmadu/service/providers/MasterKeyGenerator.java
@@ -1,5 +1,0 @@
-package com.mmadu.service.providers;
-
-public interface MasterKeyGenerator {
-    String generateMasterKey();
-}
