--- mmadu-user-service/src/main/java/com/mmadu/service/providers/MasterKeyResolver.java
@@ -1,5 +1,0 @@
-package com.mmadu.service.providers;
-
-public interface MasterKeyResolver {
-    String getMasterKey();
-}
