--- mmadu-user-service/src/main/java/com/mmadu/service/providers/NoOpKeyCipher.java
@@ -1,19 +1,0 @@
-package com.mmadu.service.providers;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.stereotype.Component;
-
-@Component
-@ConditionalOnProperty(value = "mmadu.domain.encrypt-keys", havingValue = "false", matchIfMissing = true)
-public class NoOpKeyCipher implements KeyCipher {
-
-    @Override
-    public String encrypt(String data) {
-        return data;
-    }
-
-    @Override
-    public String decrypt(String encryptedData) {
-        return encryptedData;
-    }
-}
