--- mmadu-user-service/src/test/java/com/mmadu/service/providers/DefaultPropertiesMasterKeyResolverTest.java
@@ -1,37 +1,0 @@
-package com.mmadu.service.providers;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.mockito.Mockito.doReturn;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@ContextConfiguration(classes = DefaultPropertiesMasterKeyResolver.class)
-@TestPropertySource(properties = "mmadu.security.master-key=11111111111111111111111111111111")
-public class DefaultPropertiesMasterKeyResolverTest {
-
-    @Autowired
-    private MasterKeyResolver masterKeyResolver;
-
-    @MockBean
-    private MasterKeyGenerator masterKeyGenerator;
-
-    @Before
-    public void setUp(){
-        doReturn("1232243434343434").when(masterKeyGenerator).generateMasterKey();
-    }
-
-    @Test
-    public void givenMasterKeyInPropertiesMasterKeyShouldBeTheSame() {
-        assertThat(masterKeyResolver.getMasterKey(), is(equalTo("11111111111111111111111111111111")));
-    }
-}
